Propagate I/O errors from filesystem and serialization failures
Previously, I/O and serialization/deserialization errors were silently
swallowed.

This commit makes use of the NSFileHandle APIs on iOS 13 and later to
perform filesystem operations and propagate the errors to the caller.

The following APIs are now deprecated in favor of the variants that take
a (nullable) `NSError **`:

```
-[CASQueueFile add:]
-[CASQueueFile peek:]
-[CASQueueFile pop:]
-[CASQueueFile clear]
-[CASObjectQueue add:]
-[CASObjectQueue peek:]
-[CASObjectQueue pop:]
-[CASObjectQueue clear]
```

In addition, the two convenience methods `-peek` and `-pop` to access
a single element are deprecated since there is no way to distinguish
an empty queue from an error:

```
-[CASObjectQueue peek]
-[CASObjectQueue pop]
```

Use `-peek:error:` and `-pop:error:` instead, and call `-firstObject`
on the result if you don't need to distinguish an empty queue from an
error.